+++ src/salesforce/model/salesforce-access-token-response.ts
@@ -1,5 +1,6 @@
 export interface SalesforceAccessTokenResponse {
   access_token: string;
+  refresh_token: string;
   instance_url: string;
   id: string;
   issued_at: string;