+++ src/utils/authentication/authentication-provider.ts
@@ -1,0 +1,5 @@
+export interface AuthenticationProvider<R> {
+  authenticate(): Promise<R>;
+
+  constructHeaders(): Promise<Record<string, string>>;
+}